Vibrator: Add support to offload the patterns
Extend the Vibrator HAL to support offloading the predefined
patterns to remote processor in a multi-proc system.
The applications which uses the vibrator can map their events
to the predefined patterns to maintain the same vibration effect
for an event on the remote proc after offloading.
Also,move the files as per new architecture.

Change-Id: I049ad620a2e4aba461d3dd0194329e19b5ffed18