Vibrator: set amplitude for both OL and CL vibrator
It was initially assumed that mSelectedVibrator would be determined
before setAmplitude() is called. However, this is not always the case,
and it cannot predict which vibrator path setAmplitude() will use.
To ensure no vibrator path misses the setAmplitude() call, it is
necessary to trap into both VibratorOL and VibratorCL for this API.

Change-Id: I2e352491c8e9c30b5f841ecf0dfb56896d788c14