vibratorCL: Add power wake locks
Device is going to suspend mode when vibratorcl task is waiting for dsp
callback which is resulting in conditional variable wait timeout to last
more than the expected time.
Adding power locks for vibratorCL to make sure whenever task is going
for conditional wait, acquire a lock and once cv is timeout release it
this will let device not to go in suspend state whenever a vibratorcl
task is in active state.

Change-Id: Ieb9c498a18e3873cdfa412e903445d49df8eba2c