vibratorCL: update HapticsState at start of stop stream
Binder trasaction happens from vibrator service to vibrator HAL and then it can go to vibratorOL or vibratorCL depending on the usecase triggered.
When OL usecase is triggered, CL is supposed to be stopped. From vibrator HAL CL is closed first. There is a possibility that CL is already addressing stop and in that case when we are trying to teardown CL from OL context is not supposed to be done as CL is already tearing down.
Updating HapticsState at start of stop to make sure other caller can use the flag before calling stop.

Change-Id: I787d789c564e5b1a37e8b14e46aab0734e288469